web/production/backend/reportable-app/pyproject.toml
[project]
name = "file-storage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aioboto3>=15.0.0",
    "aiofiles>=24.1.0",
    "alembic>=1.16.4",
    "fastapi>=0.116.1",
    "granian[reload]>=2.5.1",
    "openai>=2.6.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg>=3.2.9",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "redis>=6.4.0",
    "sqlalchemy>=2.0.42",
    "sqlalchemy-utils>=0.41.2",
    "types-aiobotocore[s3]>=2.24.0",
]
